src/test/java/com/leetcode/bitmanipulation/easy/PowerOfFour_342Test.java
package com.leetcode.bitmanipulation.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class PowerOfFour_342Test {
  private final PowerOfFour_342 solution = new PowerOfFour_342();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments(16, true), arguments(5, false));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void isPowerOfFour(int input, boolean expectedResult) {

    boolean actualResult = solution.isPowerOfFour(input);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
